Allow placeholders in separate postgis db config files
